Code review changes (move InputState, rename nuget pkg to _CHEMISTRY_PACKAGE_NAME
